分离出handler
